Refactor index.js to actions
